move some functions out of the main file
