+++ src/program/handle-control-c.ts
@@ -1,0 +1,8 @@
+process.on('uncaughtException', (error) => {
+    if (error instanceof Error && error.name === 'ExitPromptError') {
+        console.log('👋 until next time!')
+    } else {
+        // Rethrow unknown errors
+        throw error
+    }
+})
